Cave World : refactoring data-driven complet — blocs, textures, crafts par JSON
- IDs de blocs passent de Byte (limite 127) à Short avec familles organisées :
  100=terrain, 1000=bois, 2000=pierre, 3000=minerais, 4000=désert,
  5000=glace, 6000=liquides, 7000=végétation, 8000=fonctionnel,
  9000=cotton, 10000=items
- BlockDef + BlockRegistry : 89 JSON dans caves/blocks/, chargement dynamique,
  texture array GL de taille variable, plus aucun mapping hardcodé
- CraftDef + CraftRegistry : 14 recettes JSON dans caves/crafts/
- MeshBuilder : suppression du when(block) de 50 lignes → BlockRegistry.getLayerForFace()
- CaveRenderer : loadBlockTextures() dynamique, shader mod 128→4096
- Chunk : ByteArray→ShortArray, LodCache/LodBuilder : idem
- CaveWorldChunkStorage : format binaire Short pour les IDs de blocs
- CaveActivity : blockColor/blockTextureName/RECIPES/CREATIVE_INVENTORY supprimés
- Mobs : champ sprite ajouté dans skeleton.json et zombie.json

+++ app/src/main/java/com/Atom2Universe/app/games/caves/CaveWorldSave.kt
@@ -19,8 +19,8 @@ data class CaveWorldSave(
     var playerZ: Double,
     var playerYaw: Float,
     var playerPitch: Float,
-    var inventory: Map<Byte, Int>,
-    var hotbar: List<Byte?>,        // 9 slots, null = vide
+    var inventory: Map<Short, Int>,
+    var hotbar: List<Short?>,
     // Progression joueur
     var playerHp: Int = 20,
     var playerLevel: Int = 1,
@@ -144,12 +144,12 @@ object CaveWorldSaveManager {
 
     private fun fromJson(j: JSONObject): CaveWorldSave {
         val invJson = j.optJSONObject("inventory") ?: JSONObject()
-        val inventory = mutableMapOf<Byte, Int>()
-        invJson.keys().forEach { k -> inventory[k.toByte()] = invJson.getInt(k) }
+        val inventory = mutableMapOf<Short, Int>()
+        invJson.keys().forEach { k -> inventory[k.toShort()] = invJson.getInt(k) }
         val hotbarArr = j.optJSONArray("hotbar")
-        val hotbar: List<Byte?> = if (hotbarArr != null) {
+        val hotbar: List<Short?> = if (hotbarArr != null) {
             (0 until hotbarArr.length()).map { i ->
-                val v = hotbarArr.getInt(i); if (v < 0) null else v.toByte()
+                val v = hotbarArr.getInt(i); if (v < 0) null else v.toShort()
             }
         } else List(19) { null }
         val weaponsArr = j.optJSONArray("playerWeapons")
