+++ app/src/main/java/com/Atom2Universe/app/games/caves/world/CaveWorldChunkStorage.kt
@@ -14,42 +14,38 @@ class CaveWorldChunkStorage(worldDir: File) {
     private val executor = Executors.newSingleThreadExecutor()
 
     // Cache mémoire : clé "cx_cy_cz" → (localIndex → blockType)
-    // Chargé paresseusement depuis le disque à la première utilisation par chunk.
-    private val cache = ConcurrentHashMap<String, ConcurrentHashMap<Int, Byte>>()
+    private val cache = ConcurrentHashMap<String, ConcurrentHashMap<Int, Short>>()
 
     private fun cacheKey(cx: Int, cy: Int, cz: Int) = "${cx}_${cy}_${cz}"
     private fun diffFile(cx: Int, cy: Int, cz: Int) = File(diffsDir, "${cx}_${cy}_${cz}.diff")
 
-    // Applique les modifications du joueur sur un chunk fraîchement généré depuis la seed.
     fun applyDiff(chunk: Chunk) {
         val diff = loadDiff(chunk.cx, chunk.cy, chunk.cz) ?: return
         diff.forEach { (idx, type) -> chunk.blocks[idx] = type }
     }
 
-    // Enregistre une modification et déclenche une sauvegarde async.
-    fun recordChange(cx: Int, cy: Int, cz: Int, localIndex: Int, type: Byte) {
+    fun recordChange(cx: Int, cy: Int, cz: Int, localIndex: Int, type: Short) {
         val k = cacheKey(cx, cy, cz)
         cache.getOrPut(k) { ConcurrentHashMap() }[localIndex] = type
-        // Snapshot immutable pour l'écriture async (le cache peut continuer à muter)
         val snapshot = HashMap(cache[k]!!)
         executor.execute { writeDiff(cx, cy, cz, snapshot) }
     }
 
-    private fun loadDiff(cx: Int, cy: Int, cz: Int): Map<Int, Byte>? {
+    private fun loadDiff(cx: Int, cy: Int, cz: Int): Map<Int, Short>? {
         val k = cacheKey(cx, cy, cz)
-        cache[k]?.let { return it }   // déjà en mémoire
+        cache[k]?.let { return it }
 
         val f = diffFile(cx, cy, cz)
         if (!f.exists()) return null
 
         return runCatching {
-            val map = ConcurrentHashMap<Int, Byte>()
+            val map = ConcurrentHashMap<Int, Short>()
             GZIPInputStream(f.inputStream().buffered()).use { gz ->
                 DataInputStream(gz).use { din ->
                     val count = din.readUnsignedShort()
                     repeat(count) {
                         val idx  = din.readUnsignedShort()
-                        val type = din.readByte()
+                        val type = din.readShort()
                         map[idx] = type
                     }
                 }
@@ -59,7 +55,7 @@ class CaveWorldChunkStorage(worldDir: File) {
         }.getOrNull()
     }
 
-    private fun writeDiff(cx: Int, cy: Int, cz: Int, diff: Map<Int, Byte>) {
+    private fun writeDiff(cx: Int, cy: Int, cz: Int, diff: Map<Int, Short>) {
         runCatching {
             diffFile(cx, cy, cz).outputStream().buffered().use { os ->
                 GZIPOutputStream(os).use { gz ->
@@ -67,7 +63,7 @@ class CaveWorldChunkStorage(worldDir: File) {
                         dout.writeShort(diff.size)
                         diff.forEach { (idx, type) ->
                             dout.writeShort(idx)
-                            dout.writeByte(type.toInt())
+                            dout.writeShort(type.toInt())
                         }
                     }
                 }
